refactor(find): move support types to find file
